fix(proxy): merge middleware.ts CORS into proxy.ts; delete middleware.ts
Next.js 16 disallows having BOTH apps/broomva/middleware.ts and proxy.ts:

  Error: Both middleware file "./middleware.ts" and proxy file "./proxy.ts"
  are detected. Please use "./proxy.ts" only.

PR #187 added middleware.ts (dynamic CORS allowlist), but proxy.ts (the
auth gate) was already there. The collision broke every Vercel prod
deploy from PR #187 onward — Vercel keeps serving the last successful
pre-PR-187 build, so PR-1 (Anthropic Messages, #188), PR-2 (OpenAI Chat
Completions, #190), and the proxy /api/v1 allowlist fix (#189) are ALL
merged to main but NONE deployed. /api/v1/messages still 307s because
the new route never reached production.

Fix: fold the dynamic-CORS logic from middleware.ts into proxy.ts:
- STATIC_ALLOWLIST + isAllowedOrigin + applyCorsHeaders move into proxy.ts
- withSecurityHeaders gets req as first arg and appends CORS headers for
  /api/* requests with allowed Origin (Vary: Origin to prevent cache
  poisoning across callers)
- proxy() short-circuits OPTIONS preflight to /api/* with 204 + CORS
  before any auth check — disallowed origins get empty 204 (browser
  fails its end, standard CORS posture)
- All 6 withSecurityHeaders call sites updated to pass req

Behavior preserved from PR #187:
- broomva.tech, www.broomva.tech, broomva.github.io always allowed
- BROOMVA_CORS_EXTRA_ORIGINS env var still adds ad-hoc origins
- localhost only in non-production
- Vary: Origin prevents cache mix-ups
- PATCH method in allowlist (was missing from stale prod build)
- X-Requested-With in allow-headers

Refs: BRO-1208 parent (Edge endpoints Phase 1), unblocks BRO-1209 +
BRO-1210 + BRO-1219.
